ADD: Supervised learning example for tuning convergence
Added test script for tuning convergence based on user evaluation. The script tries to guess two numbers and the user tells 1 if it gets closer to the pair, -1 if not (bad solution), or 0 if it is correct.
Obviously, it should stop if 0, but more importantly, the optimizer gets stuck in the boundaries of the search space. Making params Real solves this issue, but decreases convergenve speed.